SessionBase: Formatting and Splitting
SessionBase: Running black and clang. Splitting up test_sessionbase.cpp tests into two files for easier maintenance.
